+++ .php-cs-fixer.dist.php
@@ -8,7 +8,7 @@ use Sirix\CsFixerConfig\ConfigBuilder;
 return ConfigBuilder::create()
     ->inDir(__DIR__ . '/src')
     ->setRules([
-        '@PHP81Migration' => true,
+        '@PHP8x1Migration' => true,
         'PedroTroller/line_break_between_method_arguments' => [
             'max-args' => 4,
             'max-length' => 140,