+++ src/Domain/Contract/EntityInterface.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Sirix\Cycle\Extension\Contract;
+namespace Sirix\Cycle\Extension\Domain\Contract;
 
 use Ramsey\Uuid\UuidInterface;
 
 interface EntityInterface
 {
-    public function getIdentifier(): null|int|UuidInterface;
+    public function getIdentifier(): int|UuidInterface|null;
 
     public function setIdentifier(int|UuidInterface $identifier): void;
 }