+++ src/Domain/Contract/RepositoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sirix\Cycle\Extension\Domain\Contract;
+
+/**
+ * @template TEntity of EntityInterface
+ */
+interface RepositoryInterface {}